requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
groq
aiofiles
langchain
langchain-groq
langchain-community
langchain-chroma
langchain-huggingface
chromadb
pypdf
sentence-transformers
